requirements.txt
neo4j
google-genai
python-dotenv
pdfplumber
